safari-600.7-branch/LayoutTests/userscripts/script-run-at-start-expected.txt

CONSOLE MESSAGE: line 1: SUCCESS
This test should log a console message on success. 
